@@ -18,37 +18,34 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 0.1.0+8
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
+  flutter: '>=3.0.0'
 
 dependencies:
+  charts_flutter: ^0.12.0
+  cupertino_icons: ^1.0.0
   flutter:
     sdk: flutter
-
-  cupertino_icons: ^1.0.0
-  charts_flutter: ^0.9.0
+  font_awesome_flutter: ^10.1.0
+  http: ^0.13.0
+  idb_sqflite: ^1.1.0
+  intl: ^0.17.0
+  package_info_plus: ^1.4.0
+  path: ^1.8.0
+  provider: ^6.0.0
   sqflite: ^2.0.0
-  sqlite3_flutter_libs: ^0.4.0
-  sqflite: ^1.3.2+1
-  sqlite3_flutter_libs: ^0.4.0
+  sqflite_common_ffi: ^2.1.0
   sqflite_wrapper:
     git:
-      url: git://github.com/tverra/sqflite-wrapper
-  font_awesome_flutter: ^8.10.0
-  shared_preferences: ^0.5.0
-  provider: ^4.3.0
-  intl: ^0.16.0
-  http: ^0.12.0
-  package_info: ^0.4.0
+      url: git@github.com:tverra/sqflite-wrapper.git
+      ref: master
+  sqlite3_flutter_libs: ^0.5.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
-  sqflite_ffi_test:
-    git:
-      url: git://github.com/tekartik/sqflite_more
-      ref: dart2
-      path: sqflite_ffi_test
+  lint: ^1.8.0
+  uuid: ^3.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
