@@ -1,9 +1,13 @@
 extension MockableDateTime on DateTime {
-  static DateTime mockTime;
+  static DateTime? mockTime;
+
   static DateTime get current {
-    if (mockTime != null)
-      return mockTime;
-    else
+    final DateTime? time = mockTime;
+
+    if (time != null) {
+      return time;
+    } else {
       return DateTime.now();
+    }
   }
 }
