@@ -1,7 +1,6 @@
-import 'dart:ui';
-
 import 'package:count_me_down/extensions.dart';
 import 'package:count_me_down/models/session.dart';
+import 'package:count_me_down/utils/data_parser.dart';
 import 'package:count_me_down/utils/percent.dart';
 import 'package:count_me_down/utils/volume.dart';
 import 'package:flutter/material.dart';
@@ -9,18 +8,18 @@ import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:intl/intl.dart';
 
 class Drink {
-  static const tableName = 'drinks';
-  static const colId = 'id';
-  static const colSessionId = 'session_id';
-  static const colName = 'name';
-  static const colVolume = 'volume';
-  static const colAlcoholConcentration = 'alcohol_concentration';
-  static const colTimestamp = 'timestamp';
-  static const colColor = 'color';
-  static const colDrinkType = 'drink_type';
-  static const relSession = 'session';
+  static const String tableName = 'drinks';
+  static const String colId = 'id';
+  static const String colSessionId = 'session_id';
+  static const String colName = 'name';
+  static const String colVolume = 'volume';
+  static const String colAlcoholConcentration = 'alcohol_concentration';
+  static const String colTimestamp = 'timestamp';
+  static const String colColor = 'color';
+  static const String colDrinkType = 'drink_type';
+  static const String relSession = 'session';
 
-  static List _drinkTypes = [
+  static const List<String> _drinkTypes = <String>[
     'beer',
     'blender',
     'cocktail',
@@ -31,17 +30,17 @@ class Drink {
     'wine_bottle',
     'wine_glass',
   ];
-  static const alcoholDensity = 0.8;
-  int id;
-  int sessionId;
-  String name;
-  Volume volume;
-  Percent alcoholConcentration;
-  DateTime timestamp;
-  Color color;
-  DrinkTypes drinkType;
+  static const double alcoholDensity = 0.8;
+  int? id;
+  int? sessionId;
+  String? name;
+  Volume? volume;
+  Percent? alcoholConcentration;
+  DateTime? timestamp;
+  Color? color;
+  DrinkTypes? drinkType;
 
-  Session session;
+  Session? session;
 
   Drink({
     this.sessionId,
@@ -54,26 +53,24 @@ class Drink {
   });
 
   Drink.fromMap(Map<String, dynamic> map) {
-    id = map[colId];
-    sessionId = map[colSessionId];
-    name = map[colName];
-    volume = map[colVolume] != null ? Volume(map[colVolume]) : null;
+    final DataParser p = DataParser();
+
+    id = p.tryParseInt(map[colId]);
+    sessionId = p.tryParseInt(map[colSessionId]);
+    name = p.tryParseString(map[colName]);
+    volume = map[colVolume] != null ? Volume(map[colVolume] as int) : null;
     alcoholConcentration = map[colAlcoholConcentration] != null
-        ? Percent(map[colAlcoholConcentration])
+        ? Percent(map[colAlcoholConcentration] as double)
         : null;
-    if (map[colTimestamp] != null) {
-      timestamp = int.tryParse(map[colTimestamp].toString()) != null
-          ? DateTime.fromMillisecondsSinceEpoch(
-              int.tryParse(map[colTimestamp].toString()),
-              isUtc: true)
-          : DateTime.tryParse(map[colTimestamp].toString());
-    }
-    color = map[colColor] != null ? Color(map[colColor]) : null;
-    drinkType =
-        map[colDrinkType] != null && _drinkTypes.contains(map[colDrinkType])
-            ? DrinkTypes.values[_drinkTypes.indexOf(map[colDrinkType])]
-            : null;
-    session = map[relSession] != null ? Session.fromMap(map[relSession]) : null;
+    timestamp = p.tryParseDateTime(map[colTimestamp]);
+    color = map[colColor] != null ? Color(map[colColor] as int) : null;
+    drinkType = map[colDrinkType] != null &&
+            _drinkTypes.contains(map[colDrinkType])
+        ? DrinkTypes.values[_drinkTypes.indexOf(map[colDrinkType] as String)]
+        : null;
+    session = map[relSession] != null
+        ? Session.fromMap(map[relSession] as Map<String, dynamic>)
+        : null;
   }
 
   static List<String> get columns {
@@ -89,68 +86,81 @@ class Drink {
     ];
   }
 
+  static List<Drink>? fromMapList(dynamic list) {
+    return (list as List<dynamic>?)
+        ?.map<Drink>(
+          (dynamic drink) => Drink.fromMap(drink as Map<String, dynamic>),
+        )
+        .toList();
+  }
+
   IconData get iconData {
     switch (drinkType) {
       case DrinkTypes.beer:
-        return FontAwesomeIcons.beer;
+        return FontAwesomeIcons.beerMugEmpty;
       case DrinkTypes.blender:
         return FontAwesomeIcons.blender;
       case DrinkTypes.cocktail:
-        return FontAwesomeIcons.cocktail;
+        return FontAwesomeIcons.martiniGlassCitrus;
       case DrinkTypes.coffee:
-        return FontAwesomeIcons.coffee;
+        return FontAwesomeIcons.mugSaucer;
       case DrinkTypes.flask:
         return FontAwesomeIcons.flask;
-      case DrinkTypes.glass_martini:
-        return FontAwesomeIcons.glassMartiniAlt;
-      case DrinkTypes.glass_whiskey:
-        return FontAwesomeIcons.glassWhiskey;
-      case DrinkTypes.wine_bottle:
+      case DrinkTypes.glassMartini:
+        return FontAwesomeIcons.martiniGlass;
+      case DrinkTypes.glassWhiskey:
+        return FontAwesomeIcons.whiskeyGlass;
+      case DrinkTypes.wineBottle:
         return FontAwesomeIcons.wineBottle;
-      case DrinkTypes.wine_glass:
-        return FontAwesomeIcons.wineGlassAlt;
+      case DrinkTypes.wineGlass:
+        return FontAwesomeIcons.wineGlassEmpty;
       default:
-        return FontAwesomeIcons.glassWhiskey;
+        return FontAwesomeIcons.whiskeyGlass;
     }
   }
 
   Map<String, dynamic> toMap({bool forQuery = false}) {
-    final Map<String, dynamic> map = <String, dynamic>{};
+    final DataParser p = DataParser(forQuery: forQuery);
+    final DrinkTypes? drinkType = this.drinkType;
 
-    map[colId] = id;
-    map[colSessionId] = sessionId;
-    map[colName] = name;
-    map[colVolume] = volume != null ? volume.millilitres : null;
-    map[colAlcoholConcentration] =
-        alcoholConcentration != null ? alcoholConcentration.fraction : null;
-    map[colTimestamp] = timestamp != null
-        ? forQuery
-            ? timestamp.millisecondsSinceEpoch
-            : timestamp.toIso8601String()
-        : null;
-    map[colColor] = color != null ? color.value : null;
-    map[colDrinkType] = drinkType != null ? _drinkTypes[drinkType.index] : null;
+    final Map<String, dynamic> map = <String, dynamic>{
+      colId: p.serializeInt(id),
+      colSessionId: p.serializeInt(sessionId),
+      colName: p.serializeString(name),
+      colVolume: volume?.millilitres,
+      colAlcoholConcentration: alcoholConcentration?.fraction,
+      colTimestamp: p.serializeDateTime(timestamp),
+      colColor: color?.value,
+      colDrinkType: drinkType != null ? _drinkTypes[drinkType.index] : null,
+    };
 
     if (!forQuery) {
-      map[relSession] =
-          session != null ? session.toMap(forQuery: forQuery) : null;
+      map.putIfAbsent(relSession, () => session?.toMap(forQuery: forQuery));
     }
 
     return map;
   }
 
   double get alcoholContentInGrams {
-    return volume.millilitres * alcoholConcentration.fraction * alcoholDensity;
+    return (volume?.millilitres ?? 0) *
+        (alcoholConcentration?.fraction ?? 0) *
+        alcoholDensity;
   }
 
   bool consumedBetween(DateTime from, DateTime to) {
+    final DateTime? timestamp = this.timestamp;
+
+    if (timestamp == null) return false;
+
     return (timestamp.isAfter(from) || timestamp.isAtSameMomentAs(from)) &&
         (timestamp.isBefore(to) || timestamp.isAtSameMomentAs(to));
   }
 
   double currentlyAbsorbedAlcohol(Duration absorptionTime) {
     final DateTime now = MockableDateTime.current;
-    final DateTime fullyAbsorbed = timestamp.add(absorptionTime);
+
+    final DateTime fullyAbsorbed =
+        (timestamp ?? DateTime.now()).add(absorptionTime);
     final bool alreadyAbsorbed = fullyAbsorbed.isBefore(now);
 
     if (alreadyAbsorbed) {
@@ -158,7 +168,8 @@ class Drink {
     }
 
     final Duration timeUntilAbsorption = fullyAbsorbed.difference(now);
-    final Duration totalAbsorptionTime = fullyAbsorbed.difference(timestamp);
+    final Duration totalAbsorptionTime =
+        fullyAbsorbed.difference(timestamp ?? DateTime.now());
     final bool absorptionCompleted = timeUntilAbsorption == Duration.zero;
     final bool absorptionNotStarted = totalAbsorptionTime < timeUntilAbsorption;
 
@@ -180,6 +191,12 @@ class Drink {
 
   @override
   String toString() {
+    final DateTime? timestamp = this.timestamp;
+
+    if (timestamp == null) {
+      return '(${volume.toString()}) - $name';
+    }
+
     return '${DateFormat('HH:mm').format(timestamp.toLocal())} '
         '(${volume.toString()}) - $name';
   }
@@ -215,8 +232,8 @@ enum DrinkTypes {
   cocktail,
   coffee,
   flask,
-  glass_martini,
-  glass_whiskey,
-  wine_bottle,
-  wine_glass,
+  glassMartini,
+  glassWhiskey,
+  wineBottle,
+  wineGlass,
 }
