@@ -1,40 +1,106 @@
-import 'dart:convert';
-
-import 'package:shared_preferences/shared_preferences.dart';
+import 'package:count_me_down/models/profile.dart';
+import 'package:count_me_down/models/session.dart';
+import 'package:count_me_down/utils/data_parser.dart';
 
 class Preferences {
-  int activeSessionId;
-  int activeProfileId;
-  String drinkWebHook;
+  static const String tableName = 'preferences';
+  static const String colId = 'id';
+  static const String colActiveSessionId = 'active_session_id';
+  static const String colActiveProfileId = 'active_profile_id';
+  static const String colDrinkWebHook = 'drink_web_hook';
+  static const String relActiveSession = 'active_session';
+  static const String relActiveProfile = 'active_profile';
+
+  int? id;
+  int? activeSessionId;
+  int? activeProfileId;
+  String? drinkWebHook;
+
+  Session? activeSession;
+  Profile? activeProfile;
 
   Preferences({this.activeSessionId, this.activeProfileId, this.drinkWebHook});
 
-  Preferences.fromJson(String json) {
-    final Map<String, dynamic> parsed = jsonDecode(json);
+  Preferences.fromMap(Map<String, dynamic> map) {
+    final DataParser p = DataParser();
 
-    activeSessionId = parsed['activeSessionId'];
-    activeProfileId = parsed['activeProfileId'];
-    drinkWebHook = parsed['drinkWebHook'];
+    id = p.tryParseInt(map[colId]);
+    activeSessionId = p.tryParseInt(map[colActiveSessionId]);
+    activeProfileId = p.tryParseInt(map[colActiveProfileId]);
+    drinkWebHook = p.tryParseString(map[colDrinkWebHook]);
+
+    activeSession = map[relActiveSession] != null
+        ? Session.fromMap(map[relActiveSession] as Map<String, dynamic>)
+        : null;
+    activeProfile = map[relActiveProfile] != null
+        ? Profile.fromMap(map[relActiveProfile] as Map<String, dynamic>)
+        : null;
   }
 
-  Preferences.initialValues() {
-    activeProfileId = 1;
+  static List<String> get columns {
+    return <String>[
+      colId,
+      colActiveSessionId,
+      colActiveProfileId,
+      colDrinkWebHook,
+    ];
   }
 
-  String toJson() {
-    final Map<String, dynamic> map = {
-      'activeSessionId': activeSessionId,
-      'activeProfileId': activeProfileId,
-      'drinkWebHook': drinkWebHook,
+  static List<Preferences>? fromMapList(dynamic list) {
+    return (list as List<dynamic>?)
+        ?.map<Preferences>(
+          (dynamic preferences) =>
+              Preferences.fromMap(preferences as Map<String, dynamic>),
+        )
+        .toList();
+  }
+
+  Map<String, dynamic> toMap({bool forQuery = false}) {
+    final DataParser p = DataParser(forQuery: forQuery);
+
+    final Map<String, dynamic> map = <String, dynamic>{
+      colId: p.serializeInt(id),
+      colActiveSessionId: p.serializeInt(activeSessionId),
+      colActiveProfileId: p.serializeInt(activeProfileId),
+      colDrinkWebHook: p.serializeString(drinkWebHook),
     };
 
-    return jsonEncode(map);
+    if (!forQuery) {
+      map.putIfAbsent(
+        relActiveSession,
+        () => activeSession?.toMap(forQuery: forQuery),
+      );
+      map.putIfAbsent(
+        relActiveProfile,
+        () => activeProfile?.toMap(forQuery: forQuery),
+      );
+    }
+
+    return map;
   }
 
-  Future<bool> save() async {
-    final SharedPreferences sharedPreferences =
-        await SharedPreferences.getInstance();
-
-    return sharedPreferences.setString('preferences', toJson());
+  Preferences copy() {
+    return Preferences.fromMap(toMap());
   }
+
+  @override
+  String toString() {
+    return toMap().toString();
+  }
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is Preferences &&
+          id == other.id &&
+          activeSessionId == other.activeSessionId &&
+          activeProfileId == other.activeProfileId &&
+          drinkWebHook == other.drinkWebHook;
+
+  @override
+  int get hashCode =>
+      id.hashCode ^
+      activeSessionId.hashCode ^
+      activeProfileId.hashCode ^
+      drinkWebHook.hashCode;
 }
