@@ -1,5 +1,6 @@
 import 'package:count_me_down/models/drink.dart';
 import 'package:count_me_down/models/profile.dart';
+import 'package:count_me_down/utils/data_parser.dart';
 
 class Session {
   static const String tableName = 'sessions';
@@ -11,14 +12,14 @@ class Session {
   static const String relProfile = 'profile';
   static const String relDrinks = 'drinks';
 
-  int id;
-  int profileId;
-  String name;
-  DateTime startedAt;
-  DateTime endedAt;
+  int? id;
+  int? profileId;
+  String? name;
+  DateTime? startedAt;
+  DateTime? endedAt;
 
-  Profile profile;
-  List<Drink> drinks;
+  Profile? profile;
+  List<Drink>? drinks;
 
   Session({
     this.profileId,
@@ -28,26 +29,20 @@ class Session {
   });
 
   Session.fromMap(Map<String, dynamic> map) {
-    id = map[colId];
-    profileId = map[colProfileId];
-    name = map[colName];
-    if (map[colStartedAt] != null) {
-      startedAt = int.tryParse(map[colStartedAt].toString()) != null
-          ? DateTime.fromMillisecondsSinceEpoch(
-              int.tryParse(map[colStartedAt].toString()),
-              isUtc: true)
-          : DateTime.tryParse(map[colStartedAt].toString());
-    }
-    if (map[colEndedAt] != null) {
-      endedAt = int.tryParse(map[colEndedAt].toString()) != null
-          ? DateTime.fromMillisecondsSinceEpoch(
-              int.tryParse(map[colEndedAt].toString()),
-              isUtc: true)
-          : DateTime.tryParse(map[colEndedAt].toString());
-    }
-    profile = map[relProfile] != null ? Profile.fromMap(map[relProfile]) : null;
+    final DataParser p = DataParser();
+
+    id = p.tryParseInt(map[colId]);
+    profileId = p.tryParseInt(map[colProfileId]);
+    name = p.tryParseString(map[colName]);
+    startedAt = p.tryParseDateTime(map[colStartedAt]);
+    endedAt = p.tryParseDateTime(map[colEndedAt]);
+    profile = map[relProfile] != null
+        ? Profile.fromMap(map[relProfile] as Map<String, dynamic>)
+        : null;
     drinks = map[relDrinks] != null
-        ? map[relDrinks].map<Drink>((d) => Drink.fromMap(d)).toList()
+        ? (map[relDrinks] as List<Map<String, dynamic>>)
+            .map<Drink>((Map<String, dynamic> d) => Drink.fromMap(d))
+            .toList()
         : null;
   }
 
@@ -61,29 +56,31 @@ class Session {
     ];
   }
 
-  Map<String, dynamic> toMap({bool forQuery = false}) {
-    final Map<String, dynamic> map = <String, dynamic>{};
+  static List<Session>? fromMapList(dynamic list) {
+    return (list as List<dynamic>?)
+        ?.map<Session>(
+          (dynamic session) => Session.fromMap(session as Map<String, dynamic>),
+        )
+        .toList();
+  }
 
-    map[colId] = id;
-    map[colProfileId] = profileId;
-    map[colName] = name;
-    map[colStartedAt] = startedAt != null
-        ? forQuery
-            ? startedAt.millisecondsSinceEpoch
-            : startedAt.toIso8601String()
-        : null;
-    map[colEndedAt] = endedAt != null
-        ? forQuery
-            ? endedAt.millisecondsSinceEpoch
-            : endedAt.toIso8601String()
-        : null;
+  Map<String, dynamic> toMap({bool forQuery = false}) {
+    final DataParser p = DataParser(forQuery: forQuery);
+
+    final Map<String, dynamic> map = <String, dynamic>{
+      colId: p.serializeInt(id),
+      colProfileId: p.serializeInt(profileId),
+      colName: p.serializeString(name),
+      colStartedAt: p.serializeDateTime(startedAt),
+      colEndedAt: p.serializeDateTime(endedAt),
+    };
 
     if (!forQuery) {
-      map[relProfile] =
-          profile != null ? profile.toMap(forQuery: forQuery) : null;
-      map[relDrinks] = drinks != null
-          ? drinks.map((d) => d.toMap(forQuery: forQuery)).toList()
-          : null;
+      map.putIfAbsent(relProfile, () => profile?.toMap(forQuery: forQuery));
+      map.putIfAbsent(
+        relDrinks,
+        () => drinks?.map((Drink d) => d.toMap(forQuery: forQuery)).toList(),
+      );
     }
 
     return map;
