@@ -1,44 +1,43 @@
+import 'package:count_me_down/database/repos/preferences_repo.dart';
 import 'package:count_me_down/models/preferences.dart';
-import 'package:count_me_down/utils/utils.dart';
+import 'package:count_me_down/utils/utils.dart' as utils;
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 class SettingsPage extends StatefulWidget {
-  static const routeName = '/settings';
+  static const String routeName = '/settings';
 
   @override
   _SettingsPageState createState() => _SettingsPageState();
 }
 
 class _SettingsPageState extends State<SettingsPage> {
-  TextEditingController _drinkHookController;
+  TextEditingController? _drinkHookController;
   bool _isLoading = false;
 
   @override
   Widget build(BuildContext context) {
     final Preferences preferences = context.watch<Preferences>();
-    if (_drinkHookController == null) {
-      _drinkHookController =
-          TextEditingController(text: preferences.drinkWebHook);
-    }
+    _drinkHookController ??=
+        TextEditingController(text: preferences.drinkWebHook);
 
     return Scaffold(
       appBar: AppBar(
-        title: Text('Settings'),
+        title: const Text('Settings'),
       ),
       body: Container(
         padding: const EdgeInsets.all(20.0),
         child: SingleChildScrollView(
           child: Column(
-            children: [
+            children: <Widget>[
               TextFormField(
                 controller: _drinkHookController,
-                decoration: InputDecoration(
+                decoration: const InputDecoration(
                   hintText: 'Drink webhook',
                   helperText: 'POST-request for when adding drinks',
                 ),
               ),
-              SizedBox(height: 40.0),
+              const SizedBox(height: 40.0),
               ElevatedButton(
                 style: ElevatedButton.styleFrom(
                   primary: Theme.of(context).primaryColor,
@@ -53,7 +52,7 @@ class _SettingsPageState extends State<SettingsPage> {
                       style: TextStyle(
                         fontSize: 17.0,
                         fontWeight: FontWeight.bold,
-                        color: Utils.getThemeTextColor(context),
+                        color: utils.getThemeTextColor(context),
                       ),
                     ),
                   ),
@@ -74,9 +73,11 @@ class _SettingsPageState extends State<SettingsPage> {
     }
 
     final Preferences preferences = context.read<Preferences>();
-    preferences.drinkWebHook = _drinkHookController.text;
-    await preferences.save();
+    preferences.drinkWebHook = _drinkHookController?.text;
+    await updatePreferences(preferences);
 
-    Navigator.of(context).pop();
+    setState(() {
+      Navigator.of(context).pop();
+    });
   }
 }
