@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 
 class MenuElement extends StatelessWidget {
-  final String text;
-  final VoidCallback onPressed;
+  final String? text;
+  final VoidCallback? onPressed;
 
-  MenuElement({this.text, this.onPressed});
+  const MenuElement({this.text, this.onPressed});
 
   @override
   Widget build(BuildContext context) {
