@@ -4,7 +4,7 @@ import 'package:count_me_down/utils/mass.dart';
 import 'package:count_me_down/utils/percent.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../test_utils.dart';
+import '../test_utils.dart' as test_utils;
 
 void main() {
   /*test('this is a test', () {
@@ -14,18 +14,18 @@ void main() {
   });*/
 
   group('fromMap', () {
-    Map<String, dynamic> _profileMap;
+    late Map<String, dynamic> _profileMap;
 
     setUp(() {
-      _profileMap = {
+      _profileMap = <String, dynamic>{
         'id': 1,
         'name': 'Profile',
         'body_weight': 75,
         'body_water_percentage': 0.6,
         'absorption_time': 3600000,
         'per_mil_metabolized_per_hour': 0.15,
-        'sessions': [
-          {
+        'sessions': <Map<String, dynamic>>[
+          <String, dynamic>{
             'id': 1,
             'profile_id': 1,
             'name': 'Session',
@@ -33,7 +33,7 @@ void main() {
             'started_at': '2021-01-30T11:55:49.291Z',
             'ended_at': '2021-01-30T12:55:49.291Z',
           },
-          {
+          <String, dynamic>{
             'id': 2,
             'profile_id': 1,
             'name': 'Session',
@@ -50,18 +50,24 @@ void main() {
 
       expect(profile.id, _profileMap['id']);
       expect(profile.name, _profileMap['name']);
-      expect(profile.bodyWeight, Mass(_profileMap['body_weight']));
-      expect(profile.bodyWaterPercentage,
-          Percent(_profileMap['body_water_percentage']));
-      expect(profile.absorptionTime,
-          Duration(milliseconds: _profileMap['absorption_time']));
-      expect(profile.perMilMetabolizedPerHour,
-          _profileMap['per_mil_metabolized_per_hour']);
-      expect(profile.sessions.length, 2);
+      expect(profile.bodyWeight, Mass(_profileMap['body_weight'] as int));
+      expect(
+        profile.bodyWaterPercentage,
+        Percent(_profileMap['body_water_percentage'] as double),
+      );
+      expect(
+        profile.absorptionTime,
+        Duration(milliseconds: _profileMap['absorption_time'] as int),
+      );
+      expect(
+        profile.perMilMetabolizedPerHour,
+        _profileMap['per_mil_metabolized_per_hour'],
+      );
+      expect(profile.sessions?.length, 2);
     });
 
     test('profile is parsed if values are null', () {
-      final Map<String, dynamic> profileMap = {
+      final Map<String, dynamic> profileMap = <String, dynamic>{
         'id': null,
         'profile_id': null,
         'name': null,
@@ -85,14 +91,14 @@ void main() {
   });
 
   group('toMap', () {
-    Profile _profile;
+    late Profile _profile;
 
     setUp(() {
       _profile = Profile(
         name: 'Profile',
         bodyWeight: Mass(75000),
         bodyWaterPercentage: Percent(0.6),
-        absorptionTime: Duration(hours: 1),
+        absorptionTime: const Duration(hours: 1),
         perMilMetabolizedPerHour: 0.15,
       )..id = 1;
 
@@ -100,14 +106,14 @@ void main() {
         Session(
           profileId: 1,
           name: 'Session',
-          startedAt: TestUtils.getDateTime(),
-          endedAt: TestUtils.getDateTime(),
+          startedAt: test_utils.getDateTime(),
+          endedAt: test_utils.getDateTime(),
         )..id = 1,
         Session(
           profileId: 1,
           name: 'Session',
-          startedAt: TestUtils.getDateTime(),
-          endedAt: TestUtils.getDateTime(),
+          startedAt: test_utils.getDateTime(),
+          endedAt: test_utils.getDateTime(),
         )..id = 2
       ];
     });
@@ -117,12 +123,16 @@ void main() {
 
       expect(profileMap['id'], _profile.id);
       expect(profileMap['name'], _profile.name);
-      expect(profileMap['body_weight'], _profile.bodyWeight.grams);
-      expect(profileMap['body_water_percentage'],
-          _profile.bodyWaterPercentage.fraction);
-      expect(profileMap['absorption_time'],
-          _profile.absorptionTime.inMilliseconds);
-      expect(profileMap['sessions'].length, 2);
+      expect(profileMap['body_weight'], _profile.bodyWeight?.grams);
+      expect(
+        profileMap['body_water_percentage'],
+        _profile.bodyWaterPercentage?.fraction,
+      );
+      expect(
+        profileMap['absorption_time'],
+        _profile.absorptionTime?.inMilliseconds,
+      );
+      expect((profileMap['sessions'] as List<Map<String, dynamic>>).length, 2);
     });
 
     test('profile is parsed if values are null', () {
@@ -144,14 +154,14 @@ void main() {
   });
 
   group('compare', () {
-    Profile _profile;
+    late Profile _profile;
 
     setUp(() {
       _profile = Profile(
         name: 'Profile',
         bodyWeight: Mass(75),
         bodyWaterPercentage: Percent(0.6),
-        absorptionTime: Duration(hours: 1),
+        absorptionTime: const Duration(hours: 1),
         perMilMetabolizedPerHour: 0.15,
       )..id = 1;
 
@@ -159,14 +169,14 @@ void main() {
         Session(
           profileId: 1,
           name: 'Session',
-          startedAt: TestUtils.getDateTime(),
-          endedAt: TestUtils.getDateTime(),
+          startedAt: test_utils.getDateTime(),
+          endedAt: test_utils.getDateTime(),
         )..id = 1,
         Session(
           profileId: 1,
           name: 'Session',
-          startedAt: TestUtils.getDateTime(),
-          endedAt: TestUtils.getDateTime(),
+          startedAt: test_utils.getDateTime(),
+          endedAt: test_utils.getDateTime(),
         )..id = 2
       ];
     });
@@ -179,7 +189,7 @@ void main() {
 
     test('objects are not equal if parameters are equal', () {
       final Profile profile = _profile.copy();
-      profile.id = profile.id + 1;
+      profile.id = profile.id! + 1;
 
       expect(false, _profile == profile);
     });
