@@ -9,7 +9,7 @@ void main() {
   });*/
 
   group('constructor with high value', () {
-    Volume _volume;
+    late Volume _volume;
 
     setUp(() {
       _volume = Volume(1000);
@@ -17,35 +17,35 @@ void main() {
 
     test('returns correct volume in millilitres', () {
       final int actual = _volume.millilitres;
-      final int expected = 1000;
+      const int expected = 1000;
 
       expect(actual, expected);
     });
 
     test('returns correct volume in centilitres', () {
       final double actual = _volume.centilitres;
-      final double expected = 100.0;
+      const double expected = 100.0;
 
       expect(actual, expected);
     });
 
     test('returns correct volume in decilitres', () {
       final double actual = _volume.decilitres;
-      final double expected = 10.0;
+      const double expected = 10.0;
 
       expect(actual, expected);
     });
 
     test('returns correct volume in litres', () {
       final double actual = _volume.litres;
-      final double expected = 1.0;
+      const double expected = 1.0;
 
       expect(actual, expected);
     });
   });
 
   group('constructor with low value', () {
-    Volume _volume;
+    late Volume _volume;
 
     setUp(() {
       _volume = Volume(1);
@@ -53,35 +53,35 @@ void main() {
 
     test('returns correct volume in millilitres', () {
       final int actual = _volume.millilitres;
-      final int expected = 1;
+      const int expected = 1;
 
       expect(actual, expected);
     });
 
     test('returns correct volume in centilitres', () {
       final double actual = _volume.centilitres;
-      final double expected = 0.1;
+      const double expected = 0.1;
 
       expect(actual, expected);
     });
 
     test('returns correct volume in decilitres', () {
       final double actual = _volume.decilitres;
-      final double expected = 0.01;
+      const double expected = 0.01;
 
       expect(actual, expected);
     });
 
     test('returns correct volume in litres', () {
       final double actual = _volume.litres;
-      final double expected = 0.001;
+      const double expected = 0.001;
 
       expect(actual, expected);
     });
   });
 
   group('constructor with negative value', () {
-    Volume _volume;
+    late Volume _volume;
 
     setUp(() {
       _volume = Volume(-1);
@@ -89,28 +89,28 @@ void main() {
 
     test('returns correct volume in millilitres', () {
       final int actual = _volume.millilitres;
-      final int expected = -1;
+      const int expected = -1;
 
       expect(actual, expected);
     });
 
     test('returns correct volume in centilitres', () {
       final double actual = _volume.centilitres;
-      final double expected = -0.1;
+      const double expected = -0.1;
 
       expect(actual, expected);
     });
 
     test('returns correct volume in decilitres', () {
       final double actual = _volume.decilitres;
-      final double expected = -0.01;
+      const double expected = -0.01;
 
       expect(actual, expected);
     });
 
     test('returns correct volume in litres', () {
       final double actual = _volume.litres;
-      final double expected = -0.001;
+      const double expected = -0.001;
 
       expect(actual, expected);
     });
@@ -120,7 +120,7 @@ void main() {
     test('litres is added', () {
       final Volume volume = Volume.exact(litres: 560);
       final int actual = volume.millilitres;
-      final int expected = 560000;
+      const int expected = 560000;
 
       expect(actual, expected);
     });
@@ -128,7 +128,7 @@ void main() {
     test('decilitres is added', () {
       final Volume volume = Volume.exact(decilitres: 560);
       final int actual = volume.millilitres;
-      final int expected = 56000;
+      const int expected = 56000;
 
       expect(actual, expected);
     });
@@ -136,7 +136,7 @@ void main() {
     test('centilitres is added', () {
       final Volume volume = Volume.exact(centilitres: 560);
       final int actual = volume.millilitres;
-      final int expected = 5600;
+      const int expected = 5600;
 
       expect(actual, expected);
     });
@@ -144,25 +144,33 @@ void main() {
     test('millilitres is added', () {
       final Volume volume = Volume.exact(millilitres: 560);
       final int actual = volume.millilitres;
-      final int expected = 560;
+      const int expected = 560;
 
       expect(actual, expected);
     });
 
     test('different units are added', () {
       final Volume volume = Volume.exact(
-          litres: 560, decilitres: 560, centilitres: 560, millilitres: 560);
+        litres: 560,
+        decilitres: 560,
+        centilitres: 560,
+        millilitres: 560,
+      );
       final int actual = volume.millilitres;
-      final int expected = 622160;
+      const int expected = 622160;
 
       expect(actual, expected);
     });
 
     test('negative values are added', () {
       final Volume volume = Volume.exact(
-          litres: -560, decilitres: -560, centilitres: -560, millilitres: -560);
+        litres: -560,
+        decilitres: -560,
+        centilitres: -560,
+        millilitres: -560,
+      );
       final int actual = volume.millilitres;
-      final int expected = -622160;
+      const int expected = -622160;
 
       expect(actual, expected);
     });
@@ -172,109 +180,73 @@ void main() {
     test('returns zero millilitres', () {
       final Volume volume = Volume(0);
 
-      expect(volume.toString(), '0 ml');
+      expect(volume.toString(), '0 mL');
     });
 
     test('returns millilitres', () {
       final Volume volume = Volume(6);
 
-      expect(volume.toString(), '6 ml');
+      expect(volume.toString(), '6 mL');
     });
 
     test('returns negative millilitres', () {
       final Volume volume = Volume(-6);
 
-      expect(volume.toString(), '-6 ml');
+      expect(volume.toString(), '−6 mL');
     });
 
     test('returns centilitres if exactly 10 millilitres', () {
       final Volume volume = Volume(10);
 
-      expect(volume.toString(), '1 cl');
+      expect(volume.toString(), '1 cL');
     });
 
     test('returns centilitres', () {
       final Volume volume = Volume(60);
 
-      expect(volume.toString(), '6 cl');
-    });
-
-    test('rounds centilitres up', () {
-      final Volume volume = Volume(15);
-
-      expect(volume.toString(), '2 cl');
-    });
-
-    test('rounds centilitres down', () {
-      final Volume volume = Volume(14);
-
-      expect(volume.toString(), '1 cl');
+      expect(volume.toString(), '6 cL');
     });
 
     test('returns negative centilitres', () {
       final Volume volume = Volume(-60);
 
-      expect(volume.toString(), '-6 cl');
+      expect(volume.toString(), '−6 cL');
     });
 
     test('returns decilitres if exactly 100 millilitres', () {
       final Volume volume = Volume(100);
 
-      expect(volume.toString(), '1 dl');
+      expect(volume.toString(), '1 dL');
     });
 
     test('returns decilitres', () {
       final Volume volume = Volume(600);
 
-      expect(volume.toString(), '6 dl');
-    });
-
-    test('rounds decilitres up', () {
-      final Volume volume = Volume(150);
-
-      expect(volume.toString(), '2 dl');
-    });
-
-    test('rounds decilitres down', () {
-      final Volume volume = Volume(149);
-
-      expect(volume.toString(), '1 dl');
+      expect(volume.toString(), '6 dL');
     });
 
     test('returns negative decilitres', () {
       final Volume volume = Volume(-600);
 
-      expect(volume.toString(), '-6 dl');
+      expect(volume.toString(), '−6 dL');
     });
 
     test('returns litres if exactly 1000 millilitres', () {
       final Volume volume = Volume(1000);
 
-      expect(volume.toString(), '1 l');
+      expect(volume.toString(), '1 L');
     });
 
     test('returns litres', () {
       final Volume volume = Volume(6000);
 
-      expect(volume.toString(), '6 l');
-    });
-
-    test('rounds litres up', () {
-      final Volume volume = Volume(1500);
-
-      expect(volume.toString(), '2 l');
-    });
-
-    test('rounds litres down', () {
-      final Volume volume = Volume(1499);
-
-      expect(volume.toString(), '1 l');
+      expect(volume.toString(), '6 L');
     });
 
     test('returns negative litres', () {
       final Volume volume = Volume(-6000);
 
-      expect(volume.toString(), '-6 l');
+      expect(volume.toString(), '−6 L');
     });
   });
 }
