@@ -1,28 +1,33 @@
+import 'package:count_me_down/database/repos/preferences_repo.dart';
 import 'package:count_me_down/models/preferences.dart';
 import 'package:count_me_down/pages/create_session_page.dart';
+import 'package:count_me_down/pages/graph_page.dart';
 import 'package:count_me_down/pages/profile_page.dart';
 import 'package:count_me_down/pages/scoreboard_page.dart';
 import 'package:count_me_down/pages/session_page.dart';
 import 'package:count_me_down/pages/sessions_page.dart';
 import 'package:count_me_down/pages/settings_page.dart';
 import 'package:count_me_down/pages/start_page.dart';
+import 'package:count_me_down/styles.dart' as styles;
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  final SharedPreferences sharedPreferences =
-      await SharedPreferences.getInstance();
-  final String preferencesJson = sharedPreferences.getString('preferences');
-  final Preferences preferences = preferencesJson != null
-      ? Preferences.fromJson(preferencesJson)
-      : Preferences.initialValues();
+  Preferences preferences = await getPreferences();
 
-  runApp(Provider(
-    create: (_) => preferences,
-    child: CountMeDownApp(),
-  ));
+  while (preferences.id == null) {
+    await Future<void>.delayed(const Duration(milliseconds: 100));
+
+    preferences = await getPreferences();
+  }
+
+  runApp(
+    Provider<Preferences>(
+      create: (_) => preferences,
+      child: CountMeDownApp(),
+    ),
+  );
 }
 
 class CountMeDownApp extends StatelessWidget {
@@ -31,25 +36,24 @@ class CountMeDownApp extends StatelessWidget {
     final Preferences preferences = context.watch<Preferences>();
 
     return MaterialApp(
-        theme: ThemeData(
-          // Tyrkisk pepper blå
-          primaryColor: Color.fromRGBO(1, 12, 142, 1),
-          // Jäger grønn
-          // primaryColor: Color.fromRGBO(11, 38, 16, 1),
-          visualDensity: VisualDensity.adaptivePlatformDensity,
-        ),
-        initialRoute: preferences.activeSessionId == null ? '/' : 'session',
-        routes: {
-          '/': (_) => StartPage(),
-          '/settings': (_) => SettingsPage(),
-          '/sessions': (_) => SessionsPage(),
-          '/sessions/createSession': (_) => CreateSessionPage(),
-          '/scoreboard': (_) => ScoreboardPage(),
-          '/profile': (_) => ProfilePage(),
-          'session': (_) => SessionPage(),
-        },
-        onGenerateRoute: (RouteSettings settings) {
-          return null;
-        });
+      theme: ThemeData(
+        primarySwatch: styles.colorToMaterialColor(styles.hotNSweetBlue),
+        appBarTheme: const AppBarTheme(backgroundColor: styles.hotNSweetBlue),
+        visualDensity: VisualDensity.adaptivePlatformDensity,
+      ),
+      initialRoute: preferences.activeSessionId == null ? '/' : 'session',
+      routes: <String, Widget Function(BuildContext)>{
+        '/': (_) => StartPage(),
+        '/settings': (_) => SettingsPage(),
+        '/sessions': (_) => SessionsPage(),
+        '/sessions/createSession': (_) => CreateSessionPage(),
+        '/scoreboard': (_) => const ScoreboardPage(),
+        '/profile': (_) => const ProfilePage(),
+        'session': (_) => const SessionPage(),
+      },
+      onGenerateRoute: (RouteSettings settings) {
+        return null;
+      },
+    );
   }
 }
