@@ -1,14 +1,14 @@
-import 'package:count_me_down/database/db_repo.dart';
+import 'package:count_me_down/database/db_repos.dart';
 import 'package:count_me_down/models/preferences.dart';
 import 'package:count_me_down/models/profile.dart';
-import 'package:count_me_down/utils/utils.dart';
+import 'package:count_me_down/utils/utils.dart' as utils;
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 class ProfileView extends StatelessWidget {
   final bool header;
 
-  ProfileView({this.header = true});
+  const ProfileView({this.header = true});
 
   @override
   Widget build(BuildContext context) {
@@ -20,51 +20,54 @@ class ProfileView extends StatelessWidget {
         right: false,
         left: false,
         bottom: false,
-        child: FutureBuilder(
-          future: ProfileRepo.getProfile(preferences.activeProfileId),
-          builder: (BuildContext context, AsyncSnapshot<Profile> snapshot) {
+        child: FutureBuilder<Profile?>(
+          future: getProfile(preferences.activeProfileId ?? 0),
+          builder: (BuildContext context, AsyncSnapshot<Profile?> snapshot) {
             if (!snapshot.hasData) {
-              return Center(
+              return const Center(
                 child: CircularProgressIndicator(),
               );
             }
 
-            final Profile profile = snapshot.data;
+            final Profile? profile = snapshot.data;
+
+            if (profile == null) return Container();
 
             return Column(
               crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                header
-                    ? Text(
-                        'Current profile:',
-                        style: TextStyle(
-                          fontSize: 20.0,
-                          fontWeight: FontWeight.bold,
-                          color: Utils.getThemeTextColor(context),
-                        ),
-                      )
-                    : Container(),
-                SizedBox(
+              children: <Widget>[
+                if (header)
+                  Text(
+                    'Current profile:',
+                    style: TextStyle(
+                      fontSize: 20.0,
+                      fontWeight: FontWeight.bold,
+                      color: utils.getThemeTextColor(context),
+                    ),
+                  )
+                else
+                  Container(),
+                const SizedBox(
                   height: 10.0,
                 ),
-                _textStyle(context, profile.name),
-                SizedBox(
+                _textStyle(context, profile.name ?? ''),
+                const SizedBox(
                   height: 5.0,
                 ),
                 _textStyle(
                   context,
                   'Weight: ',
-                  '${profile.bodyWeight.toString()}',
+                  profile.bodyWeight.toString(),
                 ),
                 _textStyle(
                   context,
                   'Body water content: ',
-                  '${profile.bodyWaterPercentage.toString()}',
+                  profile.bodyWaterPercentage.toString(),
                 ),
                 _textStyle(
                   context,
                   'Alcohol absorption time: ',
-                  '${Utils.formatDuration(profile.absorptionTime)}',
+                  utils.formatDuration(profile.absorptionTime ?? Duration.zero),
                 ),
                 _textStyle(
                   context,
@@ -79,22 +82,22 @@ class ProfileView extends StatelessWidget {
     );
   }
 
-  Widget _textStyle(BuildContext context, String leading, [String text]) {
+  Widget _textStyle(BuildContext context, String leading, [String? text]) {
     return Row(
-      children: [
+      children: <Widget>[
         Text(
-          leading ?? '',
+          leading,
           style: TextStyle(
-            color: Utils.getThemeTextColor(context),
+            color: utils.getThemeTextColor(context),
             height: 1.5,
             fontWeight: FontWeight.bold,
           ),
         ),
-        SizedBox(width: 2.0),
+        const SizedBox(width: 2.0),
         Text(
           text ?? '',
           style: TextStyle(
-            color: Utils.getThemeTextColor(context),
+            color: utils.getThemeTextColor(context),
             height: 1.5,
           ),
         ),
