@@ -1,11 +1,11 @@
-import 'alcohol_tracker_tests.dart' as alcoholTrackerTests;
-import 'mass_tests.dart' as massTests;
-import 'volume_tests.dart' as volumeTests;
-import 'percent_tests.dart' as percentTests;
+import 'alcohol_tracker_tests.dart' as alcohol_tracker_tests;
+import 'mass_tests.dart' as mass_tests;
+import 'percent_tests.dart' as percent_tests;
+import 'volume_tests.dart' as volume_tests;
 
-main() {
-  alcoholTrackerTests.main();
-  massTests.main();
-  volumeTests.main();
-  percentTests.main();
+void main() {
+  alcohol_tracker_tests.main();
+  mass_tests.main();
+  volume_tests.main();
+  percent_tests.main();
 }
